More movements, little improvments
