+++ pkg/upgrade/main.go
@@ -1,16 +1,10 @@
 package upgrade
 
 import (
-	"context"
 	"sort"
 
 	"github.com/Masterminds/semver"
-	"sigs.k8s.io/controller-runtime/pkg/client"
-
-	v1 "github.com/jaegertracing/jaeger-operator/pkg/apis/jaegertracing/v1"
 )
-
-type upgradeFunction = func(ctx context.Context, client client.Client, jaeger v1.Jaeger) (v1.Jaeger, error)
 
 var (
 	semanticVersions    []*semver.Version
@@ -46,8 +40,4 @@ func versions(versions map[string]upgradeFunction) ([]*semver.Version, error) {
 	// apply the updates in order
 	sort.Sort(semver.Collection(versionLists))
 	return versionLists, nil
-}
-
-func noop(ctx context.Context, client client.Client, jaeger v1.Jaeger) (v1.Jaeger, error) {
-	return jaeger, nil
 }
